RealMLP kernel: cast y to int64 before np.bincount
SMOKE v4 proved the 3-config preprocessing-hang fix works — both folds
trained in 2.4s + 1.6s on P100, RealMLP fit completed cleanly on GPU.
Exposed a pre-existing latent bug: pd.Series.map() yields float64 (not
int64) when the dtype promotion rules trigger, and np.bincount requires
strict int. Always killed at t+3h in prior runs before reaching that
line; now visible once preprocessing is fast.

One-line fix: .astype(np.int64) after .to_numpy() in main().

+++ kaggle_kernel/kernel_realmlp/realmlp_pytabkit.py
@@ -410,7 +410,9 @@ def main():
     train, test, orig, NUMS, CATS, combo_cols = build_features(
         train, test, orig,
     )
-    y = train[TARGET].to_numpy()
+    # map -> pandas Series is float64 if any NaN; cast to int64 so
+    # np.bincount works. fails loudly if any class label was unmapped.
+    y = train[TARGET].to_numpy().astype(np.int64)
 
     result = run_cv(train, test, orig, NUMS, CATS, combo_cols)
     oof = result["oof"]
